chore: Rename search/exhaustive/uploadstore package -> searchkv
