+++ cmd/frontend/internal/search/BUILD.bazel
@@ -32,9 +32,9 @@ go_library(
         "//internal/search",
         "//internal/search/client",
         "//internal/search/exhaustive",
+        "//internal/search/exhaustive/searchkv",
         "//internal/search/exhaustive/service",
         "//internal/search/exhaustive/store",
-        "//internal/search/exhaustive/uploadstore",
         "//internal/search/query",
         "//internal/search/result",
         "//internal/search/streaming",
