+++ cmd/worker/internal/search/BUILD.bazel
@@ -25,10 +25,10 @@ go_library(
         "//internal/observation",
         "//internal/search/client",
         "//internal/search/exhaustive",
+        "//internal/search/exhaustive/searchkv",
         "//internal/search/exhaustive/service",
         "//internal/search/exhaustive/store",
         "//internal/search/exhaustive/types",
-        "//internal/search/exhaustive/uploadstore",
         "//internal/workerutil",
         "//internal/workerutil/dbworker",
         "//internal/workerutil/dbworker/store",
