+++ internal/search/exhaustive/service/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//internal/search/client",
         "//internal/search/exhaustive/store",
         "//internal/search/exhaustive/types",
-        "//internal/search/exhaustive/uploadstore",
         "//internal/search/job",
         "//internal/search/job/jobutil",
         "//internal/search/query",
